+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0
+
+Bump oxc-parser to 0.77.2, which allows for decorators inside JS files.
+Requires Prettier >= 3.6.2
+
 ## 0.1.1
 
 Bump oxc-parser to 0.72.0. The fix the issue with code like `declare module a.b` and `declare namespace a.b`.